% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcdepth.R
\name{abcdepth}
\alias{abcdepth}
\title{Calculate ABC depth of a point}
\usage{
abcdepth(data, point, quiet = TRUE)
}
\arguments{
\item{data}{A matrix (or dataframe) with the data within which to calculate depth}

\item{point}{The point for which to calculate the depth (make sure that ncol(data) == length(point))}

\item{quiet}{Suppress java debug messages}
}
\value{
The depth of the point within the given dataset.
}
\description{
Calculate ABC depth of a point
}
